MBS-10559 extract artifactory and release conventions
