+++ build-logic/publication/src/main/kotlin/convention.publish-kotlin-library.gradle.kts
@@ -1,7 +1,8 @@
 import com.avito.android.publish.KotlinLibraryPublishExtension
 
 plugins {
-    id("convention.bintray")
+    id("convention.publish-release")
+    id("convention.publish-artifactory")
 }
 
 val publishExtension = extensions.create<KotlinLibraryPublishExtension>("publish")
